requirements.txt
numpy==1.18.5
opencv-python==4.2.0.34
pdf2image==1.13.1
Pillow==7.1.2
pkg-resources==0.0.0
pytesseract==0.3.4
typing==3.7.4.1
pysdl2==0.9.7
scipy==1.4.1